+++ test/alerts_test.py
@@ -1,0 +1,24 @@
+import unittest
+
+class AlertsTest(unittest.TestCase):
+    
+    def test_alert_list(self):
+        pass
+
+    def test_alert_following_list(self):
+        pass
+
+    def test_alert_history_list(self):
+        pass
+
+    def test_alert_history_following_list(self):
+        pass
+
+    def test_grouped_alert_list(self):
+        pass
+
+    def test_grouped_alert_following_list(self):
+        pass
+
+    def test_alert_get(self):
+        pass
