+++ test/api_key_test.py
@@ -1,0 +1,12 @@
+import unittest
+
+class ApiKeyTest(unittest.TestCase):
+
+    def test_account_api_key_list(self):
+        pass
+
+    def test_account_api_key_create(self):
+        pass
+
+    def test_account_api_key_delete(self):
+        pass
