+++ test/follower.py
@@ -1,0 +1,9 @@
+import unittest
+
+class FollowerTest(unittest.TestCase):
+
+    def test_organization_follower_list(self):
+        pass
+
+    def test_organization_follower_stop(self):
+        pass
