+++ test/follower_request.py
@@ -1,0 +1,15 @@
+import unittest
+
+class FollowerRequestTest(unittest.TestCase):
+
+    def test_organization_follower_request_list(self):
+        pass
+
+    def test_organization_follower_request_create(self):
+        pass
+
+    def test_organization_follower_request_get(self):
+        pass
+
+    def test_organization_follower_request_delete(self):
+        pass
