+++ test/following_request.py
@@ -1,0 +1,15 @@
+import unittest
+
+class FollowingRequestTest(unittest.TestCase):
+
+    def test_organization_following_request_list(self):
+        pass
+
+    def test_organization_following_request_get(self):
+        pass
+
+    def test_organization_following_request_accept(self):
+        pass
+
+    def test_organization_following_request_decline(self):
+        pass
