+++ test/following_test.py
@@ -1,0 +1,9 @@
+import unittest
+
+class FollowingTest(unittest.TestCase):
+
+    def test_organization_following_list(self):
+        pass
+
+    def test_organization_following_stop(self):
+        pass
