+++ test/invites_test.py
@@ -1,0 +1,11 @@
+import unittest
+
+class InvitesTest(unittest.TestCase):
+    def test_account_invite_list(self):
+        pass
+
+    def test_account_invite_get(self):
+        pass
+
+    def test_account_invite_accept(self):
+        pass
