+++ test/member_invite.py
@@ -1,0 +1,18 @@
+import unittest
+
+class MemberInviteTest(unittest.TestCase):
+
+    def test_organization_member_invite_list(self):
+        pass
+
+    def test_organization_member_invite_create(self):
+        pass
+
+    def test_organization_member_invite_get(self):
+        pass
+
+    def test_organization_member_invite_delete(self):
+        pass
+
+    def test_organization_member_invite_resend(self):
+        pass
