+++ test/member_test.py
@@ -1,0 +1,15 @@
+import unittest
+
+class MemberTest(unittest.TestCase):
+   
+    def test_organization_member_list(self):
+        pass
+
+    def test_organization_member_get(self):
+        pass
+
+    def test_organization_member_update(self):
+        pass
+
+    def test_organization_member_delete(self):
+        pass
