+++ test/organization_test.py
@@ -1,0 +1,12 @@
+import unittest
+
+class OrganizationTest(unittest.TestCase):
+
+    def test_organization_list(self):
+        pass
+
+    def test_organization_get(self):
+        pass
+
+    def test_organization_update(self):
+        pass
