+++ test/scan_target_test.py
@@ -1,0 +1,21 @@
+import unittest
+
+class ScanTargetTest(unittest.TestCase):
+
+    def test_organization_scan_target_list(self):
+        pass
+
+    def test_organization_scan_target_create(self):
+        pass
+
+    def test_organization_scan_target_get(self):
+        pass
+
+    def test_organization_scan_target_update(self):
+        pass
+
+    def test_organization_scan_target_delete(self):
+        pass
+
+    def test_organization_scan_target_check(self):
+        pass
