+++ test/scan_test.py
@@ -1,0 +1,15 @@
+import unittest
+
+class ScanTest(unittest.TestCase):
+
+    def test_organization_scan_target_scan_start(self):
+        pass
+
+    def test_organization_scan_target_scan_stop(self):
+        pass
+
+    def test_organization_scan_target_scan_list(self):
+        pass
+
+    def test_organization_scan_target_scan_get(self):
+        pass
