+++ test/summary_test.py
@@ -1,0 +1,15 @@
+import unittest
+
+class SummaryTest(unittest.TestCase):
+
+    def test_summary_alert(self):
+        pass
+
+    def test_summary_alert_following(self):
+        pass
+
+    def test_summary_scan(self):
+        pass
+
+    def test_summary_scan_following(self):
+        pass
