+++ test/utils_test.py
@@ -1,0 +1,11 @@
+import unittest
+
+class UtilsTest(unittest.TestCase):
+    def test_format_field(self):
+        pass
+
+    def test_output_iterable(self):
+        pass
+
+    def test_dump_json(self):
+        pass
